fbdev: msm: keep MDP core clock enabled while screen is on
Some users noted that, sometimes, the screen freezes.

I was not able to replicate the issue, so at the moment the best working
solution we have is to keep the MDP core clock enabled while the screen
is on.

This *should* not increase power consumption significantly, as the clock
was enabled anyway whenever something was moving on screen.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>