scatterlist: Disable preemption during the execution of kunmap_atomic
The kunmap_atomic(miter->addr) operation is used to unmap a memory region that was previously mapped using kmap_atomic.
Both of these functions (kmap_atomic and kunmap_atomic) are designed to be used in a non-preemptible context.

The primary characteristic of atomic mapping is that it maps pages for a short duration and immediately unmaps them to maintain atomic access to the mapped pages,
free from interruption by the scheduler.
Therefore, to maintain this atomicity during the unmapping process, it is crucial to ensure that preemption does not occur.

The WARN_ON_ONCE(preemptible()) check serves as a debugging mechanism to
catch any unexpected preemption during this critical section.