arm64: DT: msm8998: Use Qualcomm's way for LAB/IBB protection
The Sony way doesn't work correctly here in kernel 4.14, as it
will flood the SPMI bus with a continuous burst of interrupts
from LAB IRQ 0, because of bad setup.

While it is possible to fix the Sony way to detect over current,
just switch to the Qualcomm way, as this is already working fine
and gives us less things to maintain on our own.