BACKPORT: sched/uclamp: Apply modified UCLAMP support to compute_energy
Since we do not use perf domain in k4.14, there is no direct way we
can do a 1:1 application of the patch from k4.19. Instead, through
cross-examination, it is found that group_max_util essentially
refers to the calculation of the performance domain frequency in k4.19
whereas the group_norm_util is for the busy time computation. Apply
the patch according to this relation.

References:

[1]: https://github.com/Juansheng/kernel_oneplus_sm8250/commit/9a843ff48d343b24139c7edb7eac92500e733d0a

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>